[dev] aligned module separated * hyphenation included in basic techs * pipeline moved to a separate module
