+++ preprocess/utils/align.py
@@ -15,7 +15,6 @@ from preprocess.basic import (add_doc_ending_point, abbreviations,
 
 from preprocess import shallow
 from preprocess import deep
-#stechs,dtechs = {},{}
 
 from collections import OrderedDict
 
@@ -163,84 +162,3 @@ def jaccard(text1,text2):
         return 0, sentA1, sentB1
     else:
         return len(setA.intersection(setB))/float(len(setA.union(setB))), sentA1, sentB1
-
-def pipeline(text,flow=None):
-    """An easier function that allows to make a full Pipeline
-    with the subprocess that users wants. Read the restriction-
-    matrix to see what sequences of subprocess are imppossible.
-
-    Parameters
-    ----------
-    text: string to parse, generally a sentence.
-
-    steps: string list with the ordered sequence of subprocesses to
-        apply.
-
-    Returns
-    -------
-
-    parsed result : string output
-
-    TODO: make a matrix to restrict impossible sequences.
-    """
-
-    steps = OrderedDict()
-    bad_steps = []
-
-    DEFAULT_FLOW = ['replace_urls','abbreviations','expand_contractions']
-
-    #If flow is not defined do the default flow.
-    if flow is None:
-        print('Runing default pipeline')
-        return pipeline(text,DEFAULT_FLOW)
-
-    #If flow is defined check the functions
-    for step in flow:
-        if step in shallow.__techniques__:
-            steps[step]=shallow.__techniques__[step]
-        elif step in deep.__techniques__:
-            steps[step]=deep.__techniques__[step]
-        else:
-            bad_steps.append(step)
-
-    #TODO:Check if the order is possible in the matrix of possible
-    #sequences
-
-    #Apply correct functions
-    for step in steps:
-        text = steps[step](text)
-    print('The following steps are not correct:', bad_steps)
-    return text
-
-#TODO: to separate all this functions in different utils/<concept>.py
-
-def hypenation(text :str, collocations :list) -> str:
-    """Made originaly to hypenate the collocations in the original text
-    The recursive looking for collocations allow to find important 
-    expressions that define topic (of course there are better techniques
-    to do this, using Deep Learning and more complex techniques.)
-
-    Once hypenated the collocations turns into single words and are not
-    mixed with the rest. For example, if you hypenate de collocation:
-
-    [natural,language] as "natural_language" will be more informative in
-    a Luhn term evaluation than just using "natural" and "language" 
-    separately.
-
-    :param text: normalized text
-    :param collocations
-    :rtype collocations: Tuple list
-
-    """
-    for tuple in collocations:
-        expression = ''
-        replacement = ''
-        for word in tuple:
-            expression += ' ' + word
-            replacement += '_' + word
-        expression = expression.strip()
-        replacement = replacement[1:]
-
-        text = text.replace(expression,replacement)
-    return text
-
